[OJ-12496] Add bearer token authorization for Jira Server
In cases where a customer has disabled basic authentication on Jira
server (see
[here](https://confluence.atlassian.com/enterprise/disabling-basic-authentication-1044776464.html))
we want to allow using `Authorization: Bearer <token>` headers
directly instead of forcing all users to use basic auth. This is a bit
hacky (I would like to get this integrated upstream, PR for that
coming soon) but will allow customers to use bearer tokens in the
short term.

Tested this locally against a Jira server running in Docker on which I
disabled basic authentication and against the Jellyfish cloud Jira
instance, both of which worked.