feat(gitlab): Enable excluding and including repositories recursively on a group/project level
This patch adds an ability to recursively and implicitly include all repositories in a certain project when fetching data from Gitlab.

Use case: Let's say I have a main group A with 100 repositories and a subgroup B. My preferred usecase it to be able to pull data from 5 projects in group A and all projects in group B.

With the situation before this patch, I essentially had two options, either:
a) put 5 repos from group A and then X repos from subgroup B in `include_repos` git config section
b) put all _other_ repos that I'm not interested in `exclude_repos`

Both of those ways are quite suboptimal in a sense that they require a lot of manual list management.
That's why this patch introduces two new Git config parameters: `include_projects_recursively` and `exclude_projects_recursively`. Gitlab group IDs can be provided there.
The code is adjusted in such a way that it now checks not only a repo's inclusion in the include list, but also checks the repo's group inclusion in the `include_projects_recursively`. If any of these two is true, then the repo will be fetched